requirements.txt
argon2-cffi==21.1.0
Django==3.2.8
django-environ==0.4.5
django-graphql-jwt==0.3.2
django-graphql-playground==1.0
graphene-django==2.15.0
psycopg2==2.9.1
